news_sorting/requirements.txt
pickle-mixin==1.0.2
wincertstore==0.2
nltk
streamlit
scikit-learn
sklearn
Pillow